Add files via upload
Make sure to save the CSS file as styles.css in a directory named static within your project directory. The HTML templates (index.html and result.html) should be saved in a directory named templates.

With these files in place, you can run the Flask application, and users will be able to play the "Rock, Paper, Scissors" game with a colorful background.